Move BAI logic to a new module
